requirements.txt
streamlit>=1.37
pandas>=2.2
numpy>=1.26
plotly>=5.22
openpyxl>=3.1
py2neo>=2021.2.4
fastapi>=0.111
uvicorn>=0.30
pydantic>=2.7
python-dotenv>=1.0
langchain-core>=0.2
langchain-openai>=0.1
